New account panel. It allows simple email address and password changing without all the fancyness.
This account panel replaces the one from jCommunity. The jCommunity one offers too many options, and some of them are not cool for our purposes. However, this functionality is still available in jCommunity and a clever URL hacker can access them, so that will require us to remove that functionality from jCommunity (by deleting the responsible files). It is not a lot of work nor complex at all, it just means that we will have to "maintain" a "custom" version of jCommunity that is lacking a number of files. How we do this will have to be decided before deployment. One can either version a stripped jCommunity, or just document which files to delete before deployment.