Lots of changes! Most changes are towards a better structure/naming convention, which was starting to be needed. Biggest changes involve the new achmeaController base class which defines some very useful utility methods that can be used in the controllers, saving tons and tons and TONS of copied code.
Another addition that snuck its way into this commit is the facilitator's ability to now make new games.

The default controller is gone, it wasn't being used.

On the To Do list as part of the whole naming/structure thing: controller actions split up into 3 or 4 subactions to enable proper form error display (which is currently accidentally being overwritten). Expect it soon!