Companies (players) can now send messages to each other. The "document" type has been removed and a message type has been added... this is because messages are different enough from deliverables to warrant not using a single type. Deliverables are on the To Do list.
The other change in this revision is that the HTML is a little more structured. I also added HTML editor support to the configuration file but it doesn't seem to work. I posted a forum message about that, although it is unclear if we ever want to use that functionality.