Players can now see deliverables of all companies in their game.
Companies now have links to their deliverables in their show pages.

Companies now have a fancy overall score meter in their show pages. It is optimized for scores between -100 and 100.
Note: The score meter has CSS in the main CSS, but also a little in the template page. This means that if the score meter size is changed, each template with score meters will need to be edited as well. Unfortunately this is unescapable.