Tons of changes.
Created narrow and wide entries, so that all descriptions and lists now use entries (this is important because of the clear divs).

Added styled select elements.

Changes all remaining instances of je and jou into u and uw.

Fixed the styling errors with all the jCommunity pages. Also rewrote the templates and locales so that they adhere to the standard used by rest of the application (as much as possible).