Experts can now send shares. The selection of the company to send shares to is done by a temporary action: ideally this would be integrated in an index company for experts.
Did some HTML spiffyness updates to the dashboards of the expert and facilitator. I guess this is a forebode as well.

Fixed the ehader to display usable links to the expert and facilitator as well.

Fixed some typos here and there.

This version is finally a version that is fully functional again (the last few commits were a bit messy), but index company doesn't work yet (you can still click the companies if you like). It has an updated front end, it just need a functioning back end to go with that.