Moved images to directories so it's not such a huge mess.
Added total scores and total score page, and the linking structure. This will serve as a skeleton for the other two extensions as well (ideas & deliverables). The total scores page lists all rounds, and the total score page displays the total scores (share and indicators) for that round.

Note: A bug currently present is that when the game is in the stopped state, the "current" round is ill defined, and I think the last round won't be displayed. A boundary case that needs resolving.