Added database and data files.
The create.sql creates the database structure.
The insert.sql is specific to a number of Dutch game scenarios.
The .mwb file is a mysql workbench file
(http://www.mysql.com/products/workbench/).
The .xls is an excel file containing public transport data which can be
imported in the SprintCity admin panel.